fix caching logic refactor code
